Buffer frame orders on server to reduce bandwidth requirements significantly
I did some math and calculated that the sending from the server with NoDelay was creating significant TCP/IP header overhead (possibly in excess of actual data). By letting the server buffer the frame orders and sync packets (sent every tick), the overhead should now be far lower and more consistent, especially in a a game with more players. With the decreased overhead, we can increase the net tick rate to decrease network input latency.

This will increase the frequency of sync hashing, but since sync hashing needs to be done within a single frame budget anyway, it should not be a problem, may consume slightly more power. A benefit is that desyncs will be detected for the exact world frame that they occur.